fix(schema): emit double, not float, for LinkML Float mapping (KI-068)
schema/linkml.py mapped Ontolith's Float to LinkML range: float, which
real LinkML tooling treats as 32-bit xsd:float - but Ontolith's Float
is backed by Python's float (IEEE-754 double) throughout, so the
emitted range understated the actual precision. schema/rdf.py's
RDF/OWL bridge already mapped the identical value_type to XSD.double
(ADR-0036), so the two bridges described the same value differently -
ADR-0036 disclosed the mismatch on its own side, ADR-0013 didn't.

to_yaml now emits range: double. from_yaml's reverse mapping already
accepted double before this change (anticipated, just never matched
on the emission side), so this is a low-risk fix: no new parsing
logic, and the two mapping tables become consistent with each other,
not just internally self-consistent. float stays accepted on import
for backward compatibility - hand-authored LinkML and schemas
exported by a pre-KI-068 Ontolith version both still use it.

Mutation-tested: reverted the mapping, confirmed
test_float_serializes_as_double_not_float fails for the stated
reason, restored.

+++ src/ontolith/schema/linkml.py
@@ -41,7 +41,15 @@ _ValueType = Literal["Text", "Integer", "Float", "Boolean", "Date", "DateTime", 
 _VALUE_TYPE_TO_LINKML_RANGE: dict[str, str] = {
     "Text": "string",
     "Integer": "integer",
-    "Float": "float",
+    # "double" (64-bit), not "float" (32-bit) - Ontolith's Float is backed by
+    # Python's float/IEEE-754 double throughout (identical to what schema/rdf.py's
+    # RDF/OWL bridge already emits, XSD.double), so double is the round-trip-
+    # accurate choice, not a lossy 32-bit truncation on export (KI-068; ADR-0013
+    # Update, ADR-0036's own Consequences section named this exact reconciliation
+    # as future work when the RDF/OWL bridge shipped). The reverse mapping below
+    # has accepted "double" since before this change, so this makes the two
+    # tables consistent with each other, not just internally with themselves.
+    "Float": "double",
     "Boolean": "boolean",
     "Date": "date",
     "DateTime": "datetime",
@@ -52,6 +60,11 @@ _VALUE_TYPE_TO_LINKML_RANGE: dict[str, str] = {
 _LINKML_RANGE_TO_VALUE_TYPE: dict[str, str] = {
     "string": "Text",
     "integer": "Integer",
+    # "float" stays accepted on import (not just "double") for backward
+    # compatibility with hand-authored LinkML and any schema exported by an
+    # older Ontolith version before the emission side switched to "double"
+    # above (KI-068) - both losslessly represent Ontolith's own Float either
+    # way, since Python float parsing doesn't distinguish the two on read.
     "float": "Float",
     "double": "Float",
     "boolean": "Boolean",
